Moved service-specific models to their own subfolder
Moved models specific to an API service to their own subfolder as
explained in #2.
