--- src/GIDX.SDK/Models/CustomerDocumentsRequest.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace GIDX.SDK.Models
-{
-    public class CustomerDocumentsRequest : RequestBase
-    {
-        public string MerchantCustomerID { get; set; }
-    }
-}
