+++ src/GIDX.SDK/Models/DocumentLibrary/DocumentRegistrationResponse.cs
@@ -4,12 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GIDX.SDK.Models
+namespace GIDX.SDK.Models.DocumentLibrary
 {
-    public class CustomerDocumentsResponse : ResponseBase
+    public class DocumentRegistrationResponse : ResponseBase
     {
         public string MerchantCustomerID { get; set; }
-        public int DocumentCount { get; set; }
-        public List<Document> Documents { get; set; }
+        public Document Document { get; set; }
     }
 }
