+++ src/GIDX.SDK/Models/DocumentLibrary/DownloadDocumentRequest.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GIDX.SDK.Models
+namespace GIDX.SDK.Models.DocumentLibrary
 {
     public class DownloadDocumentRequest : RequestBase
     {
