+++ src/GIDX.SDK/Models/DocumentLibrary/DownloadDocumentResponse.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GIDX.SDK.Models
+namespace GIDX.SDK.Models.DocumentLibrary
 {
     public class DownloadDocumentResponse : ResponseBase
     {
